+++ bridgeService-data/src/main/java/com/adobe/campaign/tests/bridge/testdata/one/SimpleStaticMethods.java
@@ -15,8 +15,6 @@ public class SimpleStaticMethods {
 
     public static final String SUCCESS_VAL = "_Success";
 
-    //RandomManager.fetchRandomCountry(String)
-    //RandomManager.getCountries()
     public static String methodReturningString() {
         return SUCCESS_VAL;
     }
@@ -29,12 +27,10 @@ public class SimpleStaticMethods {
         return in_stringArgument + SUCCESS_VAL;
     }
 
-    //RandomManager.getRandomNumber(int)
     public static int methodAcceptingIntArgument(int in_intArgument) {
         return in_intArgument *3;
     }
 
-    //RandomManager.getUniqueEmail
     public static String methodAcceptingTwoArguments(String in_stringArgument1, String in_stringArgument2) {
         return in_stringArgument1 + "+" + in_stringArgument2 + SUCCESS_VAL;
     }
@@ -58,12 +54,25 @@ public class SimpleStaticMethods {
         return in_intArgument + SUCCESS_VAL;
     }
 
+    //For impossible Objects exception
+    public static String complexMethodAcceptor(Instantiable in_arg) {
+        return SUCCESS_VAL;
+    }
 
     //Exceptions
     //DateAndTimeTools.convertStringToDate
     public static void methodThrowingException(int in_value1, int in_value2) {
         if (in_value1==in_value2) {
             throw new IllegalArgumentException("We do not allow numbers that are equal.");
+        }
+    }
+
+    //Timeouts
+    public static void methodWithTimeOut(long in_sleepDuration) {
+        try {
+            Thread.sleep(in_sleepDuration);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
         }
     }
 }