+++ integroBridgeService/src/main/java/com/adobe/campaign/tests/bridge/service/IntegroBridgeClassLoader.java
@@ -42,7 +42,7 @@ public class IntegroBridgeClassLoader extends ClassLoader {
      *
      * @param in_classPath Full class name
      */
-    private synchronized  Class getClass(String in_classPath) {
+    private synchronized  Class getClass(String in_classPath) throws ClassNotFoundException {
 
         // is this class already loaded?
         Class cls = findLoadedClass(in_classPath);
@@ -60,10 +60,10 @@ public class IntegroBridgeClassLoader extends ClassLoader {
         // like javablogging/package/ClassToLoad.class
         String file = in_classPath.replace('.', '/')
                 + ".class";
-        byte[] b = null;
+
         try {
             // This loads the byte code data from the file
-            b = loadClassData(file);
+            byte[] b = loadClassData(file);
             // defineClass is inherited from the ClassLoader class
             // and converts the byte array into a Class
             cls = defineClass(in_classPath, b, 0, b.length);
@@ -87,7 +87,7 @@ public class IntegroBridgeClassLoader extends ClassLoader {
      *            Full class name
      */
     @Override
-    public Class loadClass(String in_classFullPath) {
+    public Class loadClass(String in_classFullPath) throws ClassNotFoundException {
         log.debug("Preparing class {}",in_classFullPath);
 
         if (isClassAmongPackagePaths(in_classFullPath)) {
@@ -112,12 +112,12 @@ public class IntegroBridgeClassLoader extends ClassLoader {
      * @throws IOException Is thrown when there
      *               was some problem reading the file
      */
-    private byte[] loadClassData(String name) throws IOException {
+    private byte[] loadClassData(String name) throws IOException, ClassNotFoundException {
         // Opening the file
         InputStream stream = getClass().getClassLoader()
                 .getResourceAsStream(name);
         if (stream==null) {
-            throw new NonExistentJavaObjectException("The given class path "+name+" could not be found.");
+            throw new ClassNotFoundException("The given class path "+name+" could not be found.");
         }
         int size = stream.available();
         byte buff[] = new byte[size];
@@ -146,7 +146,8 @@ public class IntegroBridgeClassLoader extends ClassLoader {
 
     public void setPackagePaths(String in_packagePaths) {
 
-        setPackagePaths(in_packagePaths.isEmpty() ? new HashSet<>() : new HashSet<String>(Arrays.asList(in_packagePaths.split(","))));
+        setPackagePaths(in_packagePaths.isEmpty() ? new HashSet<>() : new HashSet<>(
+                Arrays.asList(in_packagePaths.split(","))));
 
     }