+++ integroBridgeService/src/main/java/com/adobe/campaign/tests/bridge/service/JavaCalls.java
@@ -8,15 +8,16 @@
  */
 package com.adobe.campaign.tests.bridge.service;
 
-import com.adobe.campaign.tests.bridge.service.exceptions.IBSConfigurationException;
-import com.adobe.campaign.tests.bridge.service.exceptions.IBSRunTimeException;
-import com.adobe.campaign.tests.bridge.service.exceptions.NonExistentJavaObjectException;
+import com.adobe.campaign.tests.bridge.service.exceptions.*;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import java.util.*;
+import java.util.concurrent.*;
 import java.util.stream.Collectors;
 
 public class JavaCalls {
+
+    private Long timeout;
 
     private Map<String, CallContent> callContent;
 
@@ -29,6 +30,7 @@ public class JavaCalls {
         callContent = new LinkedHashMap<>();
         environmentVariables = new Properties();
         setLocalClassLoader(new IntegroBridgeClassLoader());
+        timeout = Long.parseLong(ConfigValueHandlerIBS.DEFAULT_CALL_TIMEOUT.fetchValue());
     }
 
     public Map<String, CallContent> getCallContent() {
@@ -45,10 +47,39 @@ public class JavaCalls {
      * @return A map of the results
      */
     public Object call(String in_key) {
+
         if (!this.callContent.containsKey(in_key)) {
             throw new CallDefinitionNotFoundException("Could not find a call definition with the given key "+in_key);
         }
-        return this.getCallContent().get(in_key).call(getLocalClassLoader());
+
+        Object lr_object;
+
+        ExecutorService executor = Executors.newSingleThreadExecutor();
+        Callable caller = () -> {   // Lambda Expression
+
+            return this.getCallContent().get(in_key).call(getLocalClassLoader());
+        };
+        Future<Object> future = executor.submit(caller);
+        try {
+            lr_object = getTimeout()>0 ? future.get(getTimeout(), TimeUnit.MILLISECONDS) : future.get() ;
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        } catch (ExecutionException e) {
+
+            if (e.getCause() instanceof  NonExistentJavaObjectException) {
+                throw new NonExistentJavaObjectException(e.getMessage());
+            } else if (e.getCause() instanceof AmbiguousMethodException) {
+                throw new AmbiguousMethodException(e.getMessage());
+            } else if (e.getCause() instanceof TargetJavaMethodCallException) {
+                throw new TargetJavaMethodCallException(e.getCause().getMessage(), e.getCause().getCause());
+            } else {
+                throw new IBSRunTimeException(e.getMessage());
+            }
+
+        } catch (TimeoutException e) {
+            throw new IBSTimeOutException("The call for "+in_key+" took longer than the set time limit of "+getTimeout()+"ms. Process was therefore interrupted by the Bridge Service.");
+        }
+        return lr_object;
     }
 
     /**
@@ -64,13 +95,14 @@ public class JavaCalls {
 
         JavaCallResults lr_returnObject = new JavaCallResults();
 
-        //for (String lt_key : this.getCallContent().keySet().stream().sorted().collect(Collectors.toList())) {
         for (String lt_key : this.getCallContent().keySet()) {
-            //Adapt CallContent to results
-            //this.getCallContent().get(lt_key).expandArgs(lr_returnObject);
+
+            long l_startOfCall = System.currentTimeMillis();
             Object callResult = this.call(lt_key);
+            long l_endOfCall = System.currentTimeMillis();
+
             getLocalClassLoader().getCallResultCache().put(lt_key, callResult);
-            lr_returnObject.addResult(lt_key, MetaUtils.extractValuesFromObject(callResult));
+            lr_returnObject.addResult(lt_key, MetaUtils.extractValuesFromObject(callResult), l_endOfCall - l_startOfCall);
         }
 
         return lr_returnObject;
@@ -131,4 +163,13 @@ public class JavaCalls {
     public void setLocalClassLoader(IntegroBridgeClassLoader localClassLoader) {
         this.localClassLoader = localClassLoader;
     }
+
+    public Long getTimeout() {
+        return timeout;
+    }
+
+    public void setTimeout(Long timeout) {
+        this.timeout = timeout;
+    }
+
 }